requirements.txt
flask
nltk
spacy
scikit-learn
PyPDF2
python-docx
numpy
pandas
